refactor(transport): decompose server.py via create_*_router and _prepare_request (S2)
Extract 10 inlined route handlers into routes/{health,jsonrpc,websocket,audit}.py factories; create_app is now thin wiring (_build_server_components → _wire_*). Extract _prepare_request + PreparedRequest into co-located _request_handler.py (parse→auth→envelope→timestamp→nonce gate shared by handle_message, _prepare_streaming_request, iter_websocket_stream). Route batch errors through JsonRpcErrorResponse. server.py 2256→986 LOC.
